[HGS] Migrate core gameplay code
