+++ src/lib/js/hunger_games_simulator.ts
@@ -1,7 +1,6 @@
 /* eslint-disable */
 import {Dialog, FileType, UserError} from "$lib/js/dialog";
 import {FormatError} from "$lib/js/trace";
-import default_image_src from '$lib/images/hgs-default-character.png'
 
 const light_accent = '#cea964'
 const accent_black = '#1e1e1e'
@@ -44,10 +43,6 @@ let add_character_button = document.getElementById('add-character-button-wrapper
 
 
 
-/** Clamp a number **/
-function clamp(x: number, lo: number, hi: number): number {
-    return Math.min(Math.max(x, lo), hi)
-}
 
 /** Create a data URL for an object's JSON representation. */
 function ObjectToDataURL(obj: any) {
@@ -66,26 +61,9 @@ function ParseNumberArray(str: string): number[] | null {
     return parsed
 }
 
-/** Generate a random integer in `[from;to[`. */
-function random(from: number, to: number): number {
-    return Math.floor(Math.random() * (to - from)) + from
-}
-
 /** Remove all child nodes from an element. */
 function RemoveAllChildNodes(parent: HTMLElement) {
     while (parent.firstChild) parent.removeChild(parent.firstChild)
-}
-
-/** Randomise the order of elements in an array in place. */
-function shuffle<T>(array: T[]): T[] {
-    if (array.length < 2) return array;
-    for (let i = array.length - 1; i > 0; i--) {
-        let j = random(0, i)
-        let k = array[i]
-        array[i] = array[j]
-        array[j] = k
-    }
-    return array;
 }
 
 /** Create a Blob from a string */
@@ -257,415 +235,6 @@ function RemoveAllCharacters() {
 }
 
 /// ====================================================================== ///
-///  Configuration and Versioning                                          ///
-/// ====================================================================== ///
-namespace Configuration {
-    interface MaybeConfig {version?: number}
-    interface StoredURL {url: string}
-    interface StoredBlob {data: string}
-
-    /// Legacy (unversioned) configuration.
-    namespace Legacy {
-        /// An event in a legacy config file.
-        export interface StoredEvent {
-            message: string,
-            players_involved: number,
-            fatalities: number[],
-            killers: number[],
-            enabled: boolean,
-            id: number,
-            type: string
-        }
-
-        /** The options passed to the `Tribute` constructor. */
-        interface StoredTributeOptions {
-            name: string
-            gender_select: string
-            custom_pronouns: string
-            image?: string
-        }
-
-        export type CharacterConfig = { characters: StoredTributeOptions[] }
-
-        export type Config = EventList<StoredEvent>
-
-        /** Check whether a configuration is a legacy configuration. */
-        export function is(conf: any): conf is Config { return !conf || !('version' in conf) }
-
-        /** Check whether a configuration is a legacy character configuration. */
-        export function CharacterConfigP(conf: any): conf is CharacterConfig { return !conf || !('version' in conf) }
-
-        /** Check whether two events are to be considered equal */
-        export function EventsEqual(stored: StoredEvent, event: Event) {
-            return stored && event && stored.message === event.message
-        }
-
-        /** Load an event from StoredEvent. */
-        export function LoadEvent(stored: StoredEvent, list: Event[]) {
-            let event = new Event(stored.message, stored.fatalities, stored.killers, stored.type)
-            event.enabled = stored.enabled
-            list.push(event)
-        }
-
-        /** Load characters from a file. **/
-        export async function LoadCharacters(data: CharacterConfig) {
-            const saved_characters = data.characters
-
-            saved_characters.forEach(c => AddCharacter());
-            const characters = document.getElementsByClassName('game-character')
-
-            for (let i = 0; i < characters.length; i++) {
-                const character = characters.item(i) as HTMLElement
-                const select = character.getElementsByClassName('gender-select')[0] as HTMLSelectElement
-                const name = character.getElementsByClassName('character-name')[0] as HTMLInputElement
-                const pronouns = character.getElementsByClassName('custom-gender-input')[0] as HTMLInputElement
-                const saved = saved_characters[i];
-
-                select.value = saved.gender_select;
-                name.value = saved.name
-                pronouns.value = saved.custom_pronouns
-                if (typeof saved.image === 'string') {
-                    /// Load the image from a URL if it’s not a blob. Old versions of the simulator
-                    /// would produce URLs ending in ‘[object%20Object]’ for some reason. Ignore those.
-                    if (!saved.image.startsWith('blob:') && !saved.image.endsWith('[object%20Object]')) {
-                        const img = character.getElementsByClassName('tribute-image')[0] as HTMLImageElement
-                        img.src = saved.image
-                    }
-                }
-
-                /// @ts-ignore
-                let input = select.parentElement.parentElement.getElementsByClassName('custom-gender-input-wrapper')[0] as HTMLElement
-                input.style.display = (select.value === 'other' ? 'flex' : 'none')
-            }
-        }
-    } /// namespace Legacy
-
-    /// Configuration version 1
-    export namespace V1 {
-        /// Tag info as stored in a config file.
-        export interface StoredTag {name: string}
-
-        export interface StoredTagRequirement {
-            name: string,
-            player_index: number
-        }
-
-        /// An event in a V1 config file.
-        export interface StoredEvent {
-            message: string,
-            fatalities: number[],
-            killers: number[],
-            enabled: boolean,
-            type: string,
-            tag_requirements: StoredTagRequirement[]
-        }
-
-        export interface Config {
-            readonly version: 1
-            events: EventList<StoredEvent>
-            tags: StoredTag[]
-        }
-
-        export type StoredTributeImage = StoredURL | StoredBlob
-        export interface StoredTributeOptions {
-            name: string
-            gender_select: string
-            pronoun_str: string
-            image?: StoredTributeImage
-            tags: Tag[]
-        }
-
-        export interface CharacterConfig {
-            readonly version: 1
-            characters: StoredTributeOptions[]
-        }
-
-        export const enum StoredEventTag {
-            BigLang = 'BIG LANG'
-        }
-
-        /** Check whether a configuration is a legacy configuration. */
-        export function is(conf: any): conf is Config { return conf && conf.version === 1 }
-
-        /** Check whether a configuration is a legacy character configuration. */
-        export function CharacterConfigP(conf: any): conf is CharacterConfig { return conf && conf.version === 1 }
-
-        /** Check whether two events are to be considered equal */
-        export function EventsEqual(stored: StoredEvent, event: Event) {
-            return stored !== null && event !== null
-                && stored.message === event.message
-                && ArraysEqual(stored.tag_requirements, event.requirements,
-                    (s, e) => s.name === e.tag.name && s.player_index === e.player_index)
-        }
-
-        /** Load an event from StoredEvent. */
-        export function LoadEvent(stored: StoredEvent, list: Event[]) {
-            let event = new Event(stored.message, stored.fatalities, stored.killers, stored.type)
-            event.enabled = stored.enabled && stored.type !== StoredEventTag.BigLang
-            stored.tag_requirements.forEach(t => event.require(Tag.for(t.name), t.player_index))
-            list.push(event)
-        }
-
-        /** Load a tag from a TagInfo struct */
-        export function LoadTag(data: StoredTag) { Tag.for(data.name) }
-
-        /** Stringify a tag for the purpose of storing and loading it. */
-        export function StringifyTag() {
-            /// @ts-ignore
-            const info: StoredTag = {name: this.__name}
-            return JSON.stringify(info, null, 4)
-        }
-
-        /** Check whether a tag already exists.
-         *
-         * In V1, tags consist of only a name and nothing else. Since
-         * LoadTag() calls Tag.for(), which already checks whether a tag
-         * already exists, this function can just return true.
-         */
-        export function TagExists(_: StoredTag) { return true }
-
-        /** Load characters from a file. **/
-        export async function LoadCharacters(data: CharacterConfig) {
-            const saved_characters = data.characters
-
-            saved_characters.forEach(c => AddCharacter());
-            const characters = document.getElementsByClassName('game-character')
-
-            for (let i = 0; i < characters.length; i++) {
-                const character = characters.item(i) as HTMLElement
-                const select = character.getElementsByClassName('gender-select')[0] as HTMLSelectElement
-                const name = character.getElementsByClassName('character-name')[0] as HTMLInputElement
-                const pronouns = character.getElementsByClassName('custom-gender-input')[0] as HTMLInputElement
-                const saved = saved_characters[i];
-
-                select.value = saved.gender_select;
-                name.value = saved.name
-                pronouns.value = saved.pronoun_str
-                if (typeof saved.image === 'object') {
-                    const img = character.getElementsByClassName('tribute-image')[0] as HTMLImageElement
-
-                    /// Image is a URL.
-                    if ('url' in saved.image) {
-                        /// Can’t load blob URLs, so don’t try. Old versions of the simulator would also
-                        /// produce URLs ending in ‘[object%20Object]’ for some reason. Ignore those.
-                        if (!saved.image.url.startsWith('blob:') && !saved.image.url.endsWith('[object%20Object]')) {
-                            img.src = saved.image.url
-                        }
-                    }
-
-                    /// Image is a base-64 encoded data URL.
-                    else {
-                        const blob = await fetch(saved.image.data).then(r => r.blob())
-                        img.src = URL.createObjectURL(blob)
-                    }
-                }
-
-                /// @ts-ignore
-                let input = select.parentElement.parentElement.getElementsByClassName('custom-gender-input-wrapper')[0] as HTMLElement
-                input.style.display = (select.value === 'other' ? 'flex' : 'none')
-            }
-        }
-    } /// namespace V1
-
-    type EventComparator<T> = (t: T, e: Event) => boolean
-    type EventLoader<T> = (t: T, e: Event[]) => void
-    type TagExistsP<T> = (t: T) => boolean
-    type TagLoader<T> = (t: T) => void
-
-    export const current_config_version = 1
-    export let default_config: V1.Config;
-
-    /** Check if an event exists */
-    function EventExists<T>(conf_event: T, equalp: EventComparator<T>): boolean {
-        for (const event of Game.events())
-            if (equalp(conf_event, event))
-                return true;
-        return false;
-    }
-
-    /** Load the events from the configuration. */
-    function LoadEvents<T>(lists: EventList<T>, equalp: EventComparator<T>, loader: EventLoader<T>) {
-        for (const key of Event.list_keys) {
-            /// Add the events.
-            /// @ts-ignore
-            const list = lists[key] as T[]
-            list.filter(e => !EventExists(e, equalp))
-                /// @ts-ignore
-                .forEach(e => loader(e, Game.event_lists[key]))
-        }
-    }
-
-    function LoadTags<T>(tags: T[], existsp: TagExistsP<T>, loader: TagLoader<T>) {
-        tags.filter(t => !existsp(t))
-            .forEach(t => loader(t))
-    }
-
-    /** Load a configuration. */
-    export function Load(configuration: MaybeConfig, overwrite: boolean = false, from_local_storage = false) {
-        /// Legacy file format.
-        if (Legacy.is(configuration)) {
-            if (overwrite)
-                for (const key of Event.list_keys)
-                    /// @ts-ignore
-                    Game.event_lists[key] = []
-
-            LoadEvents(configuration, Legacy.EventsEqual, Legacy.LoadEvent)
-        }
-
-        /// Versioned file format.
-        else if (V1.is(configuration)) {
-            if (overwrite) {
-                /// @ts-ignore
-                for (const key of Event.list_keys) Game.event_lists[key] = []
-                Tag.registered_tags = []
-            }
-
-            LoadTags(configuration.tags, V1.TagExists, V1.LoadTag)
-            LoadEvents(configuration.events, V1.EventsEqual, V1.LoadEvent)
-        }
-
-        /// Invalid Configuration. If we're loading from localStorage, just ignore it.
-        else if (!from_local_storage) throw Error(`Invalid config version ${configuration.version}`)
-    }
-
-    /** Load the default configuration. */
-    export function LoadDefaultConfig() { Load(Configuration.default_config, true) }
-
-    /** Create an object containing the events data to store. */
-    function SaveEvents(): EventList<V1.StoredEvent> {
-        let lists: EventList<V1.StoredEvent> = {
-            all: [],
-            bloodbath: [],
-            day: [],
-            feast: [],
-            night: [],
-        }
-
-        for (const key of Event.list_keys) {
-            /// @ts-ignore
-            const list = Game.event_lists[key] as Event[]
-            /// @ts-ignore
-            const stored_list = lists[key] as V1.StoredEvent[]
-
-            for (const e of list) {
-                /// Copy the event data.
-                let stored_event: V1.StoredEvent = {
-                    message: e.message,
-                    fatalities: e.fatalities,
-                    killers: e.killers,
-                    enabled: e.enabled,
-                    type: e.type,
-                    tag_requirements: []
-                }
-
-                /// Store the requirements.
-                for (const req of e.requirements)
-                    stored_event.tag_requirements.push({player_index: req.player_index, name: req.tag.name})
-
-                /// Add the events to the lists to store.
-                stored_list.push(stored_event)
-            }
-        }
-
-        return lists
-    }
-
-    /** Create an object containing the tag data to store. */
-    function SaveTags(): V1.StoredTag[] {
-        let tags: V1.StoredTag[] = []
-        for (const tag of Tag.registered_tags) tags.push({name: tag.name})
-        return tags
-    }
-
-    /** Save the configuration. */
-    export function Save(): V1.Config {
-        return {
-            version: current_config_version,
-            events: SaveEvents(),
-            tags: SaveTags()
-        }
-    }
-
-    /** Like the Event constructor, but creates a stored event instead. */
-    export function MakeStoredEvent(message: string, fatalities: number[] = [], killers: number[] = [],
-                                    type: string = "BUILTIN", reqs: V1.StoredTagRequirement[] = []): V1.StoredEvent {
-        return {
-            message,
-            fatalities,
-            killers,
-            type,
-            enabled: true,
-            tag_requirements: reqs
-        }
-    }
-
-    /** Save characters to a file. **/
-    export async function SaveCharacters(to_local_storage: boolean = false): Promise<V1.CharacterConfig> {
-        /// Serialise Tributes.
-        let config: V1.CharacterConfig = {
-            version: current_config_version,
-            characters: []
-        }
-
-        let lasterr
-        for (let character of document.getElementsByClassName('game-character')) {
-            try {
-                /// Image.
-                let image_src: string = (<HTMLImageElement>character.getElementsByClassName('tribute-image')[0]).src?.trim()
-                let image: V1.StoredTributeImage | undefined = undefined
-
-                /// Image is a blob. Base-64 encode it.
-                if (image_src.startsWith('blob:')) {
-                    const blob = await fetch(image_src).then(r => r.blob())
-                    const reader = new FileReader()
-                    reader.readAsDataURL(blob)
-                    image = {data: await new Promise<string>(resolve => reader.onloadend = () => resolve(reader.result as string))}
-                }
-
-                /// Image is a URL.
-                else if (image !== '') image = {url: image_src}
-
-                /// Name and gender_select value.
-                const name = (<HTMLInputElement>character.getElementsByClassName('character-name')[0]).value.trim()
-                const gender_select = (<HTMLInputElement>character.getElementsByClassName('gender-select')[0]).value
-
-                /// Pronouns.
-                const pronoun_str = (<HTMLInputElement>character.getElementsByClassName('custom-gender-input')[0]).value.trim()
-
-                /// Save the character.
-                config.characters.push({
-                    name: name,
-                    gender_select: gender_select,
-                    pronoun_str: pronoun_str,
-                    image: image,
-                    tags: [],
-                })
-            } catch (e) {
-                lasterr = e
-            }
-        }
-
-        if (lasterr && !to_local_storage) Dialog.error(`Error saving characters: ${lasterr}. Some characters may not have been saved correctly.`)
-        return config
-    }
-
-    /** Load characters from a file. **/
-    export async function LoadCharacters(data: object, from_local_storage: boolean = false) {
-        RemoveAllCharacters()
-        if (Legacy.CharacterConfigP(data)) await Legacy.LoadCharacters(data);
-        else if (V1.CharacterConfigP(data)) await V1.LoadCharacters(data);
-        else if (!from_local_storage) throw Error(`Invalid character configuration file`)
-    }
-
-    /** Default Characters **/
-    export function LoadDefaultCharacters() {
-        RemoveAllCharacters()
-        AddCharacter();
-        AddCharacter();
-    }
-} /// namespace Configuration
-
-/// ====================================================================== ///
 ///  Tags                                                                  ///
 /// ====================================================================== ///
 
@@ -779,56 +348,6 @@ class Tribute {
 
     /** Add a tag to this tribute. */
     tag(t: Tag): void { if (!this.has(t)) this.__tags.push(t) }
-}
-
-/**
- * Parse a pronoun string into a TributePronouns object.
- * @throw UserError if the string is invalid.
- */
-function ParsePronounsFromCharacterCreation(character: Element): { pronouns?: TributePronouns, uses_pronouns: boolean, plural: boolean } {
-    let option = (<HTMLInputElement>character.getElementsByClassName('gender-select')[0]).value
-
-    let plural = false
-    let uses_pronouns = true
-    let tribute_pronouns: TributePronouns | undefined = undefined
-
-    let pronoun_str
-    switch (option) {
-        case 'm':
-            pronoun_str = 'he/him/his/himself'
-            break
-        case 'f':
-            pronoun_str = 'she/her/her/herself'
-            break
-        case 'c':
-            pronoun_str = 'they/them/their/themself'
-            plural = true
-            break
-        case 'n':
-            uses_pronouns = false
-            break
-        case 'other':
-            pronoun_str = (<HTMLInputElement>character.getElementsByClassName('custom-gender-input')[0]).value.trim()
-            pronoun_str = pronoun_str.replaceAll('//', '\x1f')
-            if (!pronoun_str.match(/.+\/.+\/.+\/.+/)) throw new UserError('Custom pronouns must be of the form <code>nom/acc/gen/reflx</code>. Example: <code>they/them/their/themself</code>.')
-            break
-        default:
-            throw new UserError('Game character pronoun selection has invalid state');
-    }
-
-    if (uses_pronouns) {
-        /// @ts-ignore
-        let pronouns: string[] = pronoun_str.split('/').map(x => x.replaceAll('\x1f', '//').trim())
-        if (pronouns.includes('')) throw new UserError('Custom pronouns may not be empty! You have to specify at least one non-whitespace character for each pronoun.')
-        tribute_pronouns = {
-            nominative: pronouns[0],
-            accusative: pronouns[1],
-            genitive: pronouns[2],
-            reflexive: pronouns[3]
-        }
-    }
-
-    return {pronouns: tribute_pronouns, uses_pronouns, plural}
 }
 
 /// ====================================================================== ///
@@ -1137,63 +656,6 @@ class Game {
         this.AddEvents(Game.event_lists)
     }
 
-    /** Add all events from an event list to the game. */
-    AddEvents(event_option_list: EventList) {
-        if (event_option_list.all) {
-            for (let event_list of [this.events.bloodbath, this.events.day, this.events.night, this.events.feast])
-                event_list.push(...event_option_list.all.filter(e => e.enabled))
-        }
-
-        for (let property of [GameStage.BLOODBATH, GameStage.DAY, GameStage.NIGHT, GameStage.FEAST])
-            if (event_option_list[property])
-                this.events[property].push(...(<Event[]>event_option_list[property]).filter(e => e.enabled))
-    }
-
-    /** Determine what the next game stage should be and set it. */
-    NextGameStage() {
-        /// Start of game is always STAGE_BLOODBATH
-        if (this.round === 1) {
-            this.stage = GameStage.BLOODBATH
-            return
-        }
-
-        if (this.stage === GameStage.FEAST || this.stage === GameStage.BLOODBATH) {
-            this.days_passed++
-            this.stage = GameStage.DAY
-            return
-        }
-
-        if (this.stage === GameStage.NIGHT) {
-            /// Feast can occur before Day once every 5+ as follows
-            /// 5 rounds: 25% chance
-            /// 6 rounds: 33% chance
-            /// 7+ rounds: 50% chance
-
-            let rounds_since_feast = this.round - this.last_feast;
-
-            if (rounds_since_feast >= 5) block: {
-                if (rounds_since_feast >= 7) {
-                    if (Math.random() > .50 * (rounds_since_feast - 4)) break block
-                } else if (rounds_since_feast >= 6) {
-                    if (Math.random() > .33 * (rounds_since_feast - 4)) break block
-                } else {
-                    if (Math.random() > .25 * (rounds_since_feast - 4)) break block
-                }
-
-                this.last_feast = this.round;
-                this.stage = GameStage.FEAST
-                return
-            }
-
-            /// Otherwise, it's Day
-            this.days_passed++
-            this.stage = GameStage.DAY
-            return
-        }
-
-        this.nights_passed++
-        this.stage = GameStage.NIGHT
-    }
 
     /** Initialise 10 players on the character selection screen for debugging. */
     static PrepareDebugGame(players = 10) {
@@ -1208,159 +670,6 @@ class Game {
             this.CreateGame()
             while (this.active_game_instance) this.active_game_instance.AdvanceGame()
         }
-    }
-
-    /**
-     * Check requirements for an event.
-     *
-     * There are several restrictions as to what events we can use
-     * at any given point in time.
-     */
-    RequirementsSatisfied(event: Event, current_tribute: number, tributes_left: number, died_this_round: number): boolean {
-        /// An event must not involve more players than are left.
-        if (event.players_involved > tributes_left) return false
-
-        /// If there is a required number of fatalities, we want to drive
-        /// the number of fatalities higher. If we’ve already achieved the
-        /// required number of fatalities, we don’t want any more fatalities.
-        if (this.required_fatalities) {
-            if (died_this_round < this.required_fatalities && !event.fatalities.length) return false
-            if (died_this_round >= this.required_fatalities && event.fatalities.length) return false
-        }
-
-        /// Otherwise, if an event has fatalities, we might want to do the opposite and
-        /// reroll it depending on the fatality_reroll_rate and whether we’ve already
-        /// achieved the required number of fatalities.
-        else if (event.fatalities.length && Math.random() < this.fatality_reroll_rate) return false
-
-        /// If an event has tag requirements, it can only be chosen if the players fit those requirements.
-        for (const {tag, player_index} of event.requirements)
-            if (!this.tributes_alive[current_tribute + player_index].has(tag))
-                return false;
-
-        /// Otherwise this event is fine.
-        return true;
-    }
-
-    /**
-     * Perform the next round.
-     *
-     * This keeps choosing events randomly until all characters
-     * have acted in an event.
-     */
-    NextRound(): GameRound {
-        /// Get the number of tributes
-        let tributes_left = this.tributes_alive.length
-        let tributes_alive = tributes_left
-        let current_tribute = 0
-
-        /// Get the appropriate game stage.
-        this.NextGameStage()
-
-        let round = {
-            game_events: [],
-            tributes_dead: [],
-            number: this.round,
-            stage: this.stage
-        } as GameRound
-
-        /// Shuffle the tributes to randomise the encounters.
-        shuffle(this.tributes_alive)
-
-        let event_list: Event[]
-        switch (this.stage) {
-            case GameStage.BLOODBATH:
-                event_list = this.events.bloodbath
-                break
-            default:
-                console.error(`Unsupported stage '${this.stage}'`)
-            /// fallthrough
-            case GameStage.DAY:
-                event_list = this.events.day
-                break
-            case GameStage.NIGHT:
-                event_list = this.events.night
-                break
-            case GameStage.FEAST:
-                event_list = this.events.feast
-                break
-        }
-
-        /// If the list contains no events, then there's nothing to do.
-        if (!event_list?.length) return round
-        let died_this_round = 0
-
-        /// Randomly pick an event from the corresponding event list
-        /// whose number of tributes involved does not exceed the number
-        /// of tributes left. Ensure that every tribute is only picked once.
-        /// Repeat until no tributes are left.
-        while (tributes_left) {
-            let tributes_involved: Tribute[] = []
-            let event: Event
-
-            /// Choose an event at random. Make sure we don't fall into an infinite loop.
-            let tries = 0
-            do {
-                if (tries++ > Math.max(100, event_list.length * 10)) return round
-                event = event_list[random(0, event_list.length)]
-            } while (!this.RequirementsSatisfied(event, current_tribute, tributes_left, died_this_round))
-            tributes_left -= event.players_involved
-
-            /// Handle fatalities.
-            for (const f of event.fatalities) {
-                this.tributes_alive[current_tribute + f].died_in_round = this.round
-                round.tributes_dead.push(this.tributes_alive[current_tribute + f])
-                tributes_alive--
-                died_this_round++
-            }
-
-            for (const k of event.killers) this.tributes_alive[current_tribute + k].kills += event.fatalities.length
-
-            /// Add all players affected to the event.
-            let last = current_tribute + event.players_involved
-            for (; current_tribute < last; current_tribute++) tributes_involved.push(this.tributes_alive[current_tribute])
-
-            round.game_events.push(new GameEvent(event, tributes_involved))
-            if (tributes_alive < 2) break
-        }
-
-        /// If the user set a fixed number of deaths to happen each round, then all
-        /// of them will be at the start of the list since we first ensure that we
-        /// have the required amount. That’s fine, but we should shuffle them after
-        /// the fact so all the deaths don’t end up displayed at the top since that
-        /// can make things a bit anticlimactic.
-        shuffle(round.game_events)
-        return round;
-    }
-
-    /** Try to perform the next round. Report an error to the user if something goes wrong. */
-    SecureDoRound(): GameRound {
-        try { return this.NextRound() } catch (e) {
-            /// @ts-ignore
-            Dialog.error(FormatError(e))
-            console.log(e)
-            console.log('Game State:', this)
-            throw e
-        }
-    }
-
-    /** Display the events of the current round to the user. */
-    DisplayRoundEvents() {
-        if (this.stage === GameStage.DAY) game_title.innerHTML = `Day ${this.days_passed}`
-        else if (this.stage === GameStage.NIGHT) game_title.innerHTML = `Night ${this.nights_passed}`
-        else game_title.innerHTML = this.stage.slice(0, 1).toUpperCase() + this.stage.slice(1)
-
-        DisplayMessages(this.last_round.game_events, event_message_template, (event, message) => {
-            let image_wrapper = message.getElementsByClassName('event-message-images')[0]
-            for (const tribute of event.players_involved) image_wrapper.innerHTML += `<div class="image-wrapper">${tribute.image}</div>`
-            message.getElementsByClassName('event-message')[0].innerHTML = event.message
-        })
-
-        for (const tribute of this.last_round.tributes_dead) {
-            this.tributes_alive.splice(this.tributes_alive.indexOf(tribute), 1)
-            this.tributes_died.push(tribute)
-        }
-
     }
 
     /** Display the tributes just died this round. */
@@ -1462,152 +771,6 @@ class Game {
         (<HTMLDivElement>document.getElementById('game')).style.display = 'none'
     }
 
-    /** Determine whether the game should end based on how may tributes are alive or whether all should win. */
-    CheckGameShouldEnd() {
-        if (this.tributes_alive.length < 2 || this.all_won) {
-            this.state = GameState.END_RESULTS
-            return true
-        }
-
-        return false
-    }
-
-    /** The main state machine controlling the game. */
-    AdvanceGame() {
-        window.scrollTo(0, 0)
-
-        /// Advance the state and perform and action accordingly.
-        switch (this.state) {
-            case GameState.INITIAL:
-            case GameState.ROUND_PART_1: {
-                this.last_round = this.SecureDoRound()
-                this.DisplayRoundEvents()
-                this.rounds.push(this.last_round)
-                this.round++
-
-                if (!this.CheckGameShouldEnd())
-                    if (this.stage !== GameStage.BLOODBATH && this.stage !== GameStage.FEAST)
-                        this.state = GameState.ROUND_PART_2
-                break
-            }
-
-            case GameState.ROUND_PART_2: {
-                this.last_round = this.SecureDoRound()
-                this.DisplayRoundEvents()
-                this.rounds.push(this.last_round)
-                this.round++
-
-                if (!this.CheckGameShouldEnd()) this.state = GameState.ROUND_RESULTS
-                break
-            }
-
-            case GameState.ROUND_RESULTS: {
-                this.DisplayRoundFatalities()
-                this.state = GameState.ROUND_PART_1
-                break
-            }
-
-            case GameState.END_RESULTS: {
-                this.DisplayRoundFatalities()
-                if (!this.tributes_alive.length) this.state = GameState.END_NO_WINNER
-                else this.state = GameState.END_WINNER
-                break
-            }
-
-            case GameState.END_WINNER: {
-                this.DisplayWinners()
-                this.state = GameState.END_SUMMARY_FATALITIES
-                break
-            }
-
-            case GameState.END_NO_WINNER: {
-                this.DisplayNoWinner()
-                this.state = GameState.END_SUMMARY_FATALITIES
-                break
-            }
-
-            case GameState.END_SUMMARY_FATALITIES: {
-                this.DisplayFinalFatalities()
-                this.state = GameState.END_SUMMARY_STATS
-                break
-            }
-
-            case GameState.END_SUMMARY_STATS: {
-                this.DisplayFinalStats()
-                this.state = GameState.END
-                start_game_button.innerHTML = "End Game"
-                break
-            }
-
-            case GameState.END: {
-                this.ReturnToMainMenu()
-                Game.active_game_instance = null
-                start_game_button.innerHTML = "Proceed"
-                break
-            }
-
-            default: {
-                Dialog.error('An internal error has occurred; Game.state was ' + this.state)
-                console.error('An internal error has occurred; Game.state was ' + this.state)
-                this.state = GameState.DEAD
-                throw new Error('unreachable')
-            }
-        }
-
-        /// @ts-ignore
-        window.SetOpenImagePreview()
-    }
-
-    /** Create a new game from the current setup. */
-    static CreateGame() {
-        if (Game.active_game_instance) {
-            Dialog.error('Cannot start a new game while a game is already running. Reload the page if you want to abort your current game and start a new one.')
-            return
-        }
-
-        let tributes: Tribute[] = []
-
-        /// Get tributes
-        for (let character of document.getElementsByClassName('game-character')) {
-            /// Name
-            let name = (<HTMLInputElement>character.getElementsByClassName('character-name')[0]).value.trim()
-            if (!name) {
-                Dialog.error('One or more character names are empty!')
-                return
-            }
-
-            /// Pronouns & Number
-            let pronouns: { pronouns?: TributePronouns, uses_pronouns: boolean, plural: boolean }
-            try { pronouns = ParsePronounsFromCharacterCreation(character) } catch (e) {
-                /// @ts-ignore
-                Dialog.error(e)
-                return
-            }
-
-            /// Image
-            let src = (<HTMLImageElement>character.getElementsByClassName('tribute-image')[0]).src?.trim()
-
-            /// Add the character.
-            tributes.push(new Tribute(name, {
-                pronouns: pronouns.pronouns ?? {},
-                uses_pronouns: pronouns.uses_pronouns,
-                plural: pronouns.plural,
-                image: src ?? default_image_src,
-                tags: []
-            }))
-        }
-
-        game_content.classList.remove('flex-wrap')
-
-        Game.active_game_instance = new Game(tributes)
-
-        create_game.style.display = 'none';
-        (<HTMLDivElement>document.getElementById('game')).style.display = 'block'
-
-        Game.active_game_instance.AdvanceGame();
-        (<HTMLButtonElement>document.getElementById('advance-game')).onclick =
-            () => (<Game>Game.active_game_instance).AdvanceGame.call(Game.active_game_instance)
-    }
 
     /** Abort the current game and return to the setup screen. */
     static AbortGame() {
@@ -1616,17 +779,6 @@ class Game {
             Game.active_game_instance.state = GameState.END
             Game.active_game_instance.AdvanceGame()
         })
-    }
-
-    static events() {
-        return {
-            * [Symbol.iterator](): Iterator<Event> {
-                for (let event_list_name of Object.keys(Game.event_lists))
-                    /// @ts-ignore
-                    for (let event of Game.event_lists[event_list_name] as Event[])
-                        yield event
-            }
-        }
     }
 }
